transcriber/pyproject.toml
[project]
name = "transcriber"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aio-pika>=9.5.8",
    "db",
    "openai-whisper>=20250625",
    "pydantic-settings>=2.12.0",
    "torch>=2.10.0",
    "yt-dlp>=2026.1.29",
    "tenacity>=9.0.0",
]

[tool.uv.sources]
db = { path = "../libs/db", editable = true }
